Fix tests
Ok. So now we are printing the output when we run parser directly

e.g. ruby ./parser links.log

But, now we can also import the Parser class as a library and test it.

Now I feel safe to refactor assured that I will be alerted if I change
the external behavior.